Messing around with Collision - REMOVED
added a -NOT WORKING- move() Method that will be called every tick for the player. It should update the Position of the Player and check for Collision with the Blocky world... unfortunately it doesn't work yet, so it's commented out. I hope I can fix it soon. Also added the Camera Height via camera.getViewPos() to the block Placing/Removing funtions.