COLLISION IS WORKING :)
I discarted the appoach of using regular AABBs for the collision, because I could't get it right if more than one Block where checked for collidion. (the older versions are still commented out... but will be removed in the next clean up). The current mehtod is point based Collision detection, using all the Points making up the Bounding Box of The Player (and even more inbetween, because otherwise you could get stuck in a Block). In the Collision detection every point gets simply checked if it's inside of a block. The Future addition will be to use a BlocktypeManager to check weither the Block is passable or not.