Fixed Collision detection and Block Placing
Fixed a bug in the y-collision detection, that made the player glitch through a block, when standing close to the edge. Also improved the block placing 'radius'. Now the blocks can be palced in the radius, except for the blocks in wich the player is standing - just like it should be.