NEW: BlockTypeManger is used for creating the ChunkMesh
The BlockTypeManager/btm is now the got to method for getting Information about a Block. The Chunk stores all blocks basically as IDs in a 3D Array. When ever some information about a certain block is needed, the id is passed to the 'btm' and the requested info is returned. In the current stage this was only important for the generationof the chunk mesh, but in furutre updated this feature should should come in very handy, when creating new BlockTypes and and adding funtionalities to certain blocks.  1.BlockType: The BlockType class now contains a Stadard/SpecialBlockTexture and the info if the Block should use a single or multiple textures. 2.Chunk: thecreation ChunkmMesh is now using the Textures is the given BlockType. 3.local_player_data: Clean up of old comments; Using MIN_HEIGHT and MAX_HEIGHT now; The Player placed now the Block with id 4 (currently Cobblestone) 4.main: adding 4 simple BlockTypes to the btm.