NEW: Multiple BlockTypes can be placed  +  small fixes
The PlayerData stores now the selected BlockType in a dummy variable (will be replaced later with an inventory). Pressing the numbers from 1 to 9 change the BlockType. Chunk: the Mesh is now correctly created considering the opaqueness of the BlockTypes. Small fixes: Error Message if Texture couldn't be loaded; fixed Bug that made the player stuck when jumping up against a block.