WIP: rendering transparent Blocks
Work in Progress solution for rendering transparent Blocks. Framebuffers are used as render target for the basic, transparent and composition shader so they can be handled seperately. First all solid Blocks (and Blocks with either full or none Transparenzy per Pixel) are rendered with ALPHA TESTING. Then The transparent blocks are rendered uising OIT. Finally the two layers are put together in the compisite.fs and are displayed on the screenquad via the frame.fs

ISSUES: The calculation of the number of blocks in a Chunk needs to be fixed, so the vector memory can be preallocated correctly. the WIP Functions should be moved from main.cpp to teir own class.