ADDED: Inventory and Items
Added an Item system very similar to the block system. The ItemTypeManager (itm) stores all loaded ItemTypes and is used to acess all information about the Items. The actuall items are very simple, the just store the ID of their BlockType and the number of Items in their Stack. 9 of these Items can currently be stored in the inventory. Later a fundamental Container Class should be created as basic abstraction for the Inventory, but also for for other structures that should contain Items.