adding World class and improving Chunk loading
The World class is an inteface to load world Data, just like the config does with The settings. At the moment it's a mere placeholder and will only load/store the players Position. When Threading is implemented a Database e.g. LeveleDB might be used instead. - The inventory loading still needs to be implemented.

The Chunks will load now in a Spiral around the player, so that the chunk the player is standing in will be loaded fist.

Minor fixes: The Config setting 'WindowMaximized' sets now the SDL_WINDOW_BORDERLESS Flag instead of SDL_WINDOW_MAXIMIZED. Now the game should run in fullscreen mode without strange graphic bugs and the desktop zoom/size isn't changed any more.