Adding ALL Dependencies | small edits
All the external dependencies used to build this project (GLM, GLEW, SDL2, json.hpp). I recently encountered the problem, that I neither could build nor execute it on another Windows device. Hopefully it will work now with all the dependenies added in the 'libs/' directory. I'm not sure how to handle external libs, so I hope it's not problematic with licences ect. that I've uploaded them to this repo - Please tell me if that's acually a problem.